services/match-staging/pyproject.toml
[project]
name = "match-staging-service"
version = "0.1.0"
description = "Job Match Staging & Multi-Channel Notification Service (Vorhölle)"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pydantic>=2.10.0",
    "pyyaml>=6.0.2",
    "jinja2>=3.1.4",
    # Gmail API
    "google-api-python-client>=2.150.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.1",
    # Twilio (WhatsApp) — optional, only needed if whatsapp enabled
    # "twilio>=9.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx",  # for TestClient
    "ruff>=0.7.0",
]
whatsapp = [
    "twilio>=9.3.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
